Source/governor/requirements.txt
pymongo==3.10.1
pulsar-client==2.6.1